tests: Add tests
